requirements.txt
# ORB-15 Momentum Trading System Requirements

# Core Trading
ib-insync==0.9.86
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3

# Database
psycopg2-binary>=2.9.0

# Configuration
python-dotenv>=1.0.0

# Data Analysis
scipy>=1.10.0
matplotlib>=3.7.0
yfinance>=0.2.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Logging
coloredlogs>=15.0